+++ src/components/molecules/IconWithName/Styles.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const StyledIconWithName = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 14px;
+`;
