+++ src/components/molecules/InputText/InputText.tsx
@@ -1,6 +1,6 @@
 import { VFC, memo } from "react";
-import { index as Label } from "../atom/label/index";
-import { index as Input } from "../atom/input/index";
+import { index as Label } from "../../atom/label/index";
+import { index as Input } from "../../atom/input/index";
 
 type Props = {
   placeholder?: string;
