+++ src/components/molecules/InputTextArea/InputTextArea.tsx
@@ -1,6 +1,6 @@
 import { memo, VFC } from "react";
-import { index as Label } from "../atom/label/index";
-import { ProfileTextArea } from "../atom/textArea/ProfileTextArea";
+import { index as Label } from "../../atom/label/index";
+import { ProfileTextArea } from "../../atom/textArea/ProfileTextArea";
 
 type Props = {
   placeholder?: string;
