+++ src/components/molecules/InputWithErrorMessage/InputWithErrorMessage.tsx
@@ -1,11 +1,12 @@
 import { VFC } from "react";
-import { ErrorMsg } from "../atom/text/ErrorMsg";
-import { InputTextArea } from "../molecules/InputTextArea";
+import { ErrorMsg } from "../../atom/text/ErrorMsg";
+import { InputText } from "../../molecules/InputText/InputText";
 
 type Props = {
   placeholder: string;
   text: string;
   inputValue: string;
+  defaultValue: string;
   onChange: (
     e:
       | React.ChangeEvent<HTMLInputElement>
@@ -15,20 +16,22 @@ type Props = {
   errorMessage: string;
 };
 
-export const TextAreaWithErrorMessage: VFC<Props> = ({
+export const InputWithErrorMessage: VFC<Props> = ({
   placeholder,
   text,
   inputValue,
+  defaultValue,
   onChange,
   isValid,
   errorMessage,
 }) => {
   return (
     <>
-      <InputTextArea
+      <InputText
         placeholder={placeholder}
         text={text}
         inputValue={inputValue}
+        defaultValue={defaultValue}
         onChange={onChange}
       />
       <ErrorMsg isValid={isValid}>{errorMessage}</ErrorMsg>
