+++ src/components/molecules/LogoWithTypography/Styles.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const StyledLogoContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+`;
