+++ src/components/molecules/NameWithTimestamp/NameWithTimestamp.tsx
@@ -1,9 +1,14 @@
 import { VFC } from "react";
-import styled from "styled-components";
 
-import { DEVICE } from "../../Themes/Device";
-import { index as Typography } from "../atom/typography/index";
-import LikedIcon from "../../images/liked.png";
+import { index as Typography } from "../../atom/typography/index";
+import LikedIcon from "../../../images/liked.png";
+
+import {
+  StyledNameWithTimestamp,
+  StyledTimeAndLike,
+  StyledLiked,
+  StyledLikeButton,
+} from "./Styles";
 
 type Props = {
   username?: string;
@@ -31,30 +36,3 @@ export const NameWithTimestamp: VFC<Props> = ({
     </StyledNameWithTimestamp>
   );
 };
-
-const StyledTimeAndLike = styled.div`
-  display: flex;
-  margin-top: 2px;
-`;
-
-const StyledNameWithTimestamp = styled.div`
-  margin-left: 0.4rem;
-
-  @media ${DEVICE.laptop} {
-    margin-left: 0.8rem;
-  }
-`;
-
-const StyledLikeButton = styled.img`
-  width: 12px;
-  height: 12px;
-  max-width: 12px;
-  max-height: 12px;
-  margin-right: 2px;
-`;
-
-const StyledLiked = styled.div`
-  display: flex;
-  align-items: center;
-  margin-left: 5px;
-`;
