+++ src/components/molecules/PostButtons/PostButtons.tsx
@@ -1,13 +1,13 @@
 import { VFC } from "react";
-import styled from "styled-components";
 
-import { AddImageIcon } from "../atom/icon/AddImageIcon";
-import { PreviewIcon } from "../atom/icon/PreviewIcon";
-import { QuestionIcon } from "../atom/icon/QuestionIcon";
-import { DraftIcon } from "../atom/icon/DraftIcon";
-import { SendIcon } from "../atom/icon/SendIcon";
-import { isValidPost } from "../../Themes/Validations";
-import { DEVICE } from "../../Themes/Device";
+import { AddImageIcon } from "../../atom/icon/AddImageIcon";
+import { PreviewIcon } from "../../atom/icon/PreviewIcon";
+import { QuestionIcon } from "../../atom/icon/QuestionIcon";
+import { DraftIcon } from "../../atom/icon/DraftIcon";
+import { SendIcon } from "../../atom/icon/SendIcon";
+import { isValidPost } from "../../../Themes/Validations";
+
+import { StyledButtonWrap } from "./Styles";
 
 type Props = {
   title: string;
@@ -63,22 +63,3 @@ export const PostButtons: VFC<Props> = ({
     </StyledButtonWrap>
   );
 };
-
-const StyledButtonWrap = styled.div`
-  position: fixed;
-  bottom: 20px;
-  display: flex;
-  justify-content: flex-end;
-  gap: 0 20px;
-  margin: 0 auto;
-  padding: 20px 0;
-
-  @media ${DEVICE.laptopL} {
-    top: 160px;
-    right: 3%;
-    bottom: auto;
-    justify-content: flex-start;
-    flex-direction: column;
-    gap: 14px 0;
-  }
-`;
