+++ src/components/molecules/PostInputArea/PostInputArea.tsx
@@ -1,14 +1,18 @@
 import { VFC } from "react";
-import styled from "styled-components";
-import { IoIosCloseCircle } from "react-icons/io";
 
-import { MarkdownTextArea } from "../atom/textArea/MarkdownTextArea";
-import { ReactMarkdownArea } from "../atom/reactMarkdown/index";
-import { TitleTextArea } from "../atom/textArea/TitleTextArea";
+import { MarkdownTextArea } from "../../atom/textArea/MarkdownTextArea";
+import { ReactMarkdownArea } from "../../atom/reactMarkdown/index";
+import { TitleTextArea } from "../../atom/textArea/TitleTextArea";
+import { checkLabelLength } from "../../../Themes/Validations";
 
-import { checkLabelLength } from "../../Themes/Validations";
-import { DEVICE } from "../../Themes/Device";
-import { COLOR } from "../../Themes/Color";
+import {
+  StyledTextAreaWrap,
+  StyledLabels,
+  StyledLabel,
+  StyledIoIosCloseCircle,
+  StyledInputLabelArea,
+  StyledInput,
+} from "./Styles";
 
 type Props = {
   title: string;
@@ -80,66 +84,3 @@ export const PostInputArea: VFC<Props> = ({
     </>
   );
 };
-
-const StyledTextAreaWrap = styled.div`
-  display: flex;
-  max-width: 90vw;
-  overflow: hidden;
-  border-radius: 10px;
-
-  @media ${DEVICE.laptopL} {
-    justify-content: space-between;
-    max-width: 1200px;
-  }
-  @media ${DEVICE.desktop} {
-    max-width: 1800px;
-  }
-`;
-
-const StyledLabels = styled.ul`
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  gap: 4px;
-  margin-bottom: 20px;
-`;
-
-const StyledLabel = styled.li`
-  font-size: 12px;
-  background-color: ${COLOR.GRAY};
-  color: ${COLOR.WHITE};
-  padding: 6px 8px;
-  display: inline-block;
-  border-radius: 4px;
-`;
-
-const StyledIoIosCloseCircle = styled(IoIosCloseCircle)`
-  font-size: 14px;
-  margin-left: 4px;
-
-  &:hover {
-    cursor: pointer;
-  }
-`;
-
-const StyledInputLabelArea = styled.div`
-  display: flex;
-  gap: 10px;
-`;
-
-const StyledInput = styled.input`
-  border-radius: 4px;
-  border: 1px solid rgba(0, 0, 0, 24%);
-  padding: 4px 10px;
-  margin-bottom: 10px;
-  display: block;
-  width: 400px;
-  font-size: 14px;
-  outline: none;
-  box-sizing: border-box;
-  caret-color: ${COLOR.PRIMARY};
-
-  &:focus {
-    border: 1px solid ${COLOR.PRIMARY};
-  }
-`;
