+++ src/components/molecules/SidebarUserProfile/SidebarUserProfile.tsx
@@ -1,11 +1,11 @@
 import { VFC } from "react";
-import styled from "styled-components";
 
-import { index as Icon } from "../atom/icon/index";
-import { index as Typography } from "../atom/typography";
-import { index as Link } from "../atom/link/index";
-import { SocialIcons } from "./SocialIcons";
-import { COLOR } from "../../Themes/Color";
+import { index as Icon } from "../../atom/icon/index";
+import { index as Typography } from "../../atom/typography";
+import { index as Link } from "../../atom/link/index";
+import { SocialIcons } from "../SocialIcons/SocialIcons";
+
+import { StyledSidebarUserProfile, StyledSidebarAuthor } from "./Styles";
 
 type Props = {
   uid: string;
@@ -60,21 +60,3 @@ export const SidebarUserProfile: VFC<Props> = ({
     </StyledSidebarUserProfile>
   );
 };
-
-const StyledSidebarUserProfile = styled.div`
-  width: 100%;
-  background-color: ${COLOR.WHITE};
-  border-radius: 10px;
-  display: grid;
-  gap: 12px;
-  padding: 20px;
-  box-sizing: border-box;
-`;
-
-const StyledSidebarAuthor = styled.div`
-  display: flex;
-  justify-content: flex-start;
-  gap: 14px;
-  align-items: center;
-  max-width: 220px;
-`;
