+++ src/components/molecules/SocialIcons/SocialIcons.tsx
@@ -1,10 +1,13 @@
 import { VFC } from "react";
-import { AiFillGithub, AiOutlineLink } from "react-icons/ai";
-import { FaTwitter } from "react-icons/fa";
-import styled from "styled-components";
 
-import { COLOR } from "../../Themes/Color";
-import { DEVICE } from "../../Themes/Device";
+import {
+  StyledLinkLists,
+  StyledLinkItems,
+  StyledGitHubIcon,
+  StyledTwitterIcon,
+  StyledBlogIcon,
+  StyledLinkText,
+} from "./Styles";
 
 type Props = {
   githubName: string;
@@ -73,72 +76,3 @@ export const SocialIcons: VFC<Props> = ({
     </StyledLinkLists>
   );
 };
-
-const StyledLinkLists = styled.ul`
-  display: flex;
-`;
-const StyledLinkItems = styled.li`
-  position: relative;
-  line-height: 1;
-
-  &:not(:last-child) {
-    margin-right: 10px;
-  }
-
-  &:hover {
-    cursor: pointer;
-    color: ${COLOR.BLACK};
-  }
-`;
-
-const StyledLinkText = styled.p`
-  position: absolute;
-  bottom: -20px;
-  left: 50%;
-  transform: translateX(-50%);
-  background-color: ${COLOR.GRAY};
-  color: ${COLOR.WHITE};
-  font-size: 12px;
-  text-align: center;
-  border-radius: 4px;
-  padding: 2px 20px;
-`;
-
-const StyledGitHubIcon = styled(AiFillGithub)`
-  font-size: 24px;
-  color: ${COLOR.BLACK};
-
-  @media ${DEVICE.tabletL} {
-    font-size: 26px;
-
-    &:hover {
-      color: ${COLOR.PRIMARY};
-    }
-  }
-`;
-
-const StyledTwitterIcon = styled(FaTwitter)`
-  font-size: 24px;
-  color: ${COLOR.BLACK};
-
-  @media ${DEVICE.tabletL} {
-    font-size: 26px;
-
-    &:hover {
-      color: ${COLOR.PRIMARY};
-    }
-  }
-`;
-
-const StyledBlogIcon = styled(AiOutlineLink)`
-  font-size: 24px;
-  color: ${COLOR.BLACK};
-
-  @media ${DEVICE.tabletL} {
-    font-size: 26px;
-
-    &:hover {
-      color: ${COLOR.PRIMARY};
-    }
-  }
-`;
