+++ src/components/organisms/Sidebar/Presenter.tsx
@@ -2,8 +2,8 @@ import { VFC } from "react";
 import { TwitterShareButton, TwitterIcon } from "react-share";
 
 import { index as Typography } from "../../atom/typography/index";
-import { SidebarLabels } from "../../molecules/SidebarLabels";
-import { SidebarUserProfile } from "../../molecules/SidebarUserProfile";
+import { SidebarLabels } from "../../molecules/SidebarLabels/SidebarLabels";
+import { SidebarUserProfile } from "../../molecules/SidebarUserProfile/SidebarUserProfile";
 import noLike from "../../../images/no-like.png";
 import LikedIcon from "../../../images/liked.png";
 import { COLOR } from "../../../Themes/Color";
